% tinyrox says don't edit this manually, but it can't stop you!
\name{print.pensar_research}
\alias{print.pensar_research}
\title{Print a pensar research session result}
\usage{
\method{print}{pensar_research}(x, ...)
}
\arguments{
\item{x}{A \code{pensar_research} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a pensar research session result
}
